First runthrough of entire site to new architecture (empty no-arg constructors, adf removed, etc).  Still needs a lot of testing.

+++ source/WEB-INF/classes/com/dneero/formbeans/BloggerIndex.java
@@ -42,26 +42,27 @@ public class BloggerIndex implements Serializable {
         } else {
             showmarketingmaterial = false;
         }
-
-        List<Responsepending> responsependings = HibernateUtil.getSession().createCriteria(Responsepending.class)
-                               .add(Restrictions.eq("userid", Jsf.getUserSession().getUser().getUserid()))
-                               .setCacheable(true)
-                               .list();
-        for (Iterator<Responsepending> iterator = responsependings.iterator(); iterator.hasNext();) {
-            Responsepending responsepending = iterator.next();
-            userhasresponsependings = true;
-            //If we have a blogger at this point, commit the pending response
-            if (Jsf.getUserSession().getUser().getBloggerid()>0){
-                Survey survey = Survey.get(responsepending.getSurveyid());
-                try{
-                    SurveyResponseParser srp = new SurveyResponseParser(responsepending.getResponseasstring());
-                    BloggerSurveyTake.createResponse(survey,  srp, Blogger.get(Jsf.getUserSession().getUser().getBloggerid()), responsepending.getReferredbyblogid());
-                    responsependingmsg = responsependingmsg + "You just earned $"+ Str.formatForMoney(survey.getWillingtopayperrespondent())+"! We have successfully committed your response to '"+survey.getTitle()+"'!  But don't forget to post this survey to your blog to earn even more money... click <a href='bloggersurveyposttoblog.jsf?surveyid="+survey.getSurveyid()+"'>here</a>." + "<br/><br/>";
-                } catch (ComponentException cex){
-                    responsependingmsg = responsependingmsg + "There was an error committing your response to the survey '"+survey.getTitle()+"': " + cex.getErrorsAsSingleString() + "  But don't worry... we're always adding <a href='/blogger/bloggersurveylist.jsf'>new survey opportunities</a>.<br/><br/>";
+        if (Jsf.getUserSession()!=null && Jsf.getUserSession().getUser()!=null){
+            List<Responsepending> responsependings = HibernateUtil.getSession().createCriteria(Responsepending.class)
+                                   .add(Restrictions.eq("userid", Jsf.getUserSession().getUser().getUserid()))
+                                   .setCacheable(true)
+                                   .list();
+            for (Iterator<Responsepending> iterator = responsependings.iterator(); iterator.hasNext();) {
+                Responsepending responsepending = iterator.next();
+                userhasresponsependings = true;
+                //If we have a blogger at this point, commit the pending response
+                if (Jsf.getUserSession().getUser().getBloggerid()>0){
+                    Survey survey = Survey.get(responsepending.getSurveyid());
+                    try{
+                        SurveyResponseParser srp = new SurveyResponseParser(responsepending.getResponseasstring());
+                        BloggerSurveyTake.createResponse(survey,  srp, Blogger.get(Jsf.getUserSession().getUser().getBloggerid()), responsepending.getReferredbyblogid());
+                        responsependingmsg = responsependingmsg + "You just earned $"+ Str.formatForMoney(survey.getWillingtopayperrespondent())+"! We have successfully committed your response to '"+survey.getTitle()+"'!  But don't forget to post this survey to your blog to earn even more money... click <a href='bloggersurveyposttoblog.jsf?surveyid="+survey.getSurveyid()+"'>here</a>." + "<br/><br/>";
+                    } catch (ComponentException cex){
+                        responsependingmsg = responsependingmsg + "There was an error committing your response to the survey '"+survey.getTitle()+"': " + cex.getErrorsAsSingleString() + "  But don't worry... we're always adding <a href='/blogger/bloggersurveylist.jsf'>new survey opportunities</a>.<br/><br/>";
+                    }
+                    //Delete the responsepending, now that it's been handled
+                    responsepending.delete();
                 }
-                //Delete the responsepending, now that it's been handled
-                responsepending.delete();
             }
         }
     }
